SPLICE-962 Create split per file in WholeTextInputFormat
Move test out of platform directory, it is generic
